requirements.txt
django==5.2.12
whitenoise==6.12.0
djangorestframework==3.17.1
django-cors-headers==4.9.0
python-dotenv==1.2.2
google-genai==1.70.0
googlemaps==4.10.0
google-cloud-texttospeech==2.36.0
weasyprint==68.1
Pillow==12.2.0
requests==2.33.1
celery==5.6.3
redis==7.4.0
